requirements/base.txt
SQLAlchemy~=1.4
psycopg2-binary~=2.9
python-dotenv~=0.21
pydantic~=1.10.2
alembic~=1.8
